fix(plugin): move buildHiddenAgentConfig out of the entry module — opencode 1.17 invokes every entry export as a plugin factory
opencode 1.17 calls every exported function in a plugin's entry module
(index.ts) as its own plugin factory: fn(ctx). The helper buildHiddenAgentConfig
was exported from index.ts (introduced upstream in 29a49cb6 'D19 resilience
safe parts'), so opencode invoked it as buildHiddenAgentConfig(ctx) — passing
the plugin context as `prompt` and undefined as `allowedTools` — which threw
'undefined is not an object (evaluating allowedTools)' during plugin load.
Result: no hooks/tools registered, all ctx_* tools dead, empty MC runtime log,
migration churn each boot.

The entry module must export ONLY default (the plugin factory). Move the helper
to a sibling module (hidden-agent-config.ts) where it can still be exported and
unit-tested without being mis-invoked as a factory. index-refresh.test.ts now
imports it from there.

Verified: dist export keys = ['default']; full plugin suite green (the lone
tui-config ordering flake is pre-existing on master); tsc + lint clean.
